Upgrades k8s to v1.25.13 + updates related components (#852)
* Updates k8s to v1.25.13 + update related components

This also takes the opportunity to update some other workloads like
cAdvisor, kube-state-metrics and Reloader.

* Downgrades Kured Helm chart version to 5.1.0

v5.2.0 adds a --concurrency flag to the argument list, which does not
exist in the version of Kured we are using. We are not using the latest
version of Kured because the latest only guarantees k8s compatability
back to v1.26.x and we are only on v1.25. It would probably work just
fine, but just being safe.