Updates ndt-canary DS for GCP, not Equinix
The ndt-canary DaemonSet was previously configured to work for Equinix
Metal machines, while we were testing those in production. That
experiment finished and the Metal machine deleted.

Now, we want to test MIGs in production. This commit simply copied
ndt-virtual.jsonnet over ndt-canary.jsonnet, then modified
ndt-canary.jsonnet to be appropriate for testing MIGs in production.