Updates kube-rbac-proxy in node-exporter pod (#858)
We just had an issue where the node-exporter pod on mlab1-bom01 wasn't
able to download the kube-rbac-proxy image from quay.io. The quay.io
status page said that they were aware of an intermittent issue with
downloading older images. The version of kube-rbac-proxy we were using
in the node-exporter pod was more than 2 years old. This commit updates
kube-rbac-proxy to the latest version, and while I was at it I bumped
the version of node-exporter itself, which wasn't all that old... might
as well keep everything up to date when touching something.