Adds new ndt-server -label flag for "managed"
There is a new metadata file at /var/local/metadata/managed, which contains
different values depending on the deployment environment. For "full" physical
sites it will be "switch,machine", for "minimal" sites it will be "machine" and
for BYOS it will be "none". This will allow data users to filter based on the
environment.