Adds new mlab/managed=switch-machine nodeSelector to Disco (#877)
Today, Disco is configured to run as a DaemonSet on all physical machines.
With the advent of "minimal" physical sites which do not have a switch, we
don't want Disco running there. This commit adds a new nodeSelector leveraging
the new "[mlab/managed](https://github.com/m-lab/epoxy-images/pull/261)" node
label such that Disco should only be deployed on physical machines where a
switch is present.