Reduces terminationGracePeriodSeconds from 180s to 100s
The 180s termination grace period was to account for the very long time it
could take for mlab-ns to notice that pod was shutting down. With Locate and
Heartbeat, this process only takes 10s. Nothing changes about the configuration
here, except removing unnecessary wait time. Tests still have 30s to complete,
and pusher still has 60s to push data to GCS as fast as it can.

Resolves https://github.com/m-lab/ops-tracker/issues/1742