Removes the "processes" collector from node_exporter
For a long time, logs on platform machines have been completely swamped/filled
with AppArmor error messages related to node_exporter for containerd's default
AppArmor profile. This makes it harder to debug on a machine when the logs are
so full of useless messages. This commit removes the processes collector from
node_exporter, which is only used by two dashboards and isn't really necessary.