Condenses flannel to a single DaemonSet (#902)
* Reduces flannel to a single DaemonSet

For many years we have run two flannel DaemonSets, once which deploys to
physical machines and one to virtual machines. This was all because we use
multus CNI on physical machine, but don't on virtual machines, so needed to
write a separate CNI config for each machine type to /etc/cni/net.d/.

This commit condenses flannel into a single DaemonSet that runs on everything,
and CNI config files are baked into boot images instead of being written by the
flannel pods.

This reduces a good deal of duplication and complexity in our configs, and
those CNI config files basically never changes, so baking them into images,
while less flexible, makes the overall config more simple and easy to
understand and work with.

* Updates system.jsonnet to reflect single flannel DaemonSet

* Adds missing comma to flannel.jsonnet