Make Prometheus in revtr-sidecar listen on a cluster-local address (#906)
* Makes revtr-sidecar Prom listen on private cluster address

* Temporarily removes PRIVATE_IP from prometheus.port flag

The flag name is literal, and expects only an int. I've left the PRIVATE_IP env
variable, but removed it from the flag. I submitted a PR to the revtr-sidecar
repository to address this:

https://github.com/NEU-SNS/revtr-sidecar/pull/8

Once that PR is merged, assuming it does get merged, then I will update this
again to specify a full address instead of just a port.

* Updates revtr-sidecar to v1.4.2

This includes an update to revtr-sidecar which changes the flag
-prometheus.port to -prometheus.addr, and allows us to cause Prometheus to
listen on a cluster-local address instead of a public interface.