Adds support for MultiNetworkPolicy definitions (#907)
* Adds support for MultiNetworkPolicy objects

We want to be able to apply NetworkPolicy definitions to experiment pods, but
your CNI plugin needs to support applying NetworkPolicy definitions. We use
both flannel and multus, either of which support this. However, there is
another effort from the k8s Network Plumbing Working Group (same people who
maintain multus) to implement a system that allows NetworkPolicy definitions to
pods with multiple interfaces.

https://github.com/k8snetworkplumbingwg/multi-networkpolicy
https://github.com/k8snetworkplumbingwg/multi-networkpolicy-iptables

* Condenses multi-networkpolicy configmaps into a single one

* Fixes location of closing bracket in multi-networkpolicy configmap

* Fixes multi-networkplicy DaemonSet and add test policy for NDT

* Changes experiment net-attach-defs to CNI v0.3.1

It was previously v0.2.0, which is what index2ip was designed around, but I
discovered that using that version was impacting multus' functionality with
regard to annotating pods correctly, and possibly in other ways we weren't even
noticing.

* Use v1beta2 of MultiNetworkPolicy

This version introduces the "endPort" field, allowing you to specify a port
range, which we need to do for ndt5.

Also, allow additional ports for ndt: 3001, 3010, and ephemeral port range. The
range reflects the value of net.ipv4.ip_local_port_range on our Ubuntu systems.

* Removes namespaceSelector from MultiNetworkPolicy

It was not valid in that location.

* Uses image from m-lab fork of multi-networkpolicy-iptables

Also updates ICMP flags to use "accept" instead of "allow"

* Adds MultiNetworkPolicy defs for msak, revt, wehe and neubot

* Modifies the allowed ports for revtr in its MultiNetworkPolicy

* Loads MultiNetworkPolicies for mask, revtr, wehe and neubot

* Adds podSelectors for MultiNetworkPolicies

After adding multiple MultiNetworkPolicies the multi-networkpolicy DaemonSets
seems to not be applying the rules properly. I had figured that the
NetworkAttachmentDefinition would be sufficient. This may or may not change/fix
anything.

* Formats MultiNetworkPolicies as jsonnet

I had accidentally inserted some plain JSON.

* Removes podSelector from MultiNetworkPolicies

It turns out the issue was apparently just me pushing changes to k8s-support in
sandbox with the multi-networkpolicy changes, and then another branch without,
and then with, and then without, which got pod annotations all messed up.

* Removes sample iptables custom rules from configmap

The rules are not utilized by multi-networkpolicy unless you pass special
flags, which are outlined in a comment in the file.

* Adds new MultiNetworkPolicy template to templates.jsonnet

* Adds a MultiNetworkPolicy definition to the ndt.jsonnet manifest

* Adds MultiNetworkPolicy definition to revtr manifest

* Adds MultiNetworkPolicy definition to the wehe manifest

* Adds MultiNetworkPolicy definition to msak manifest

* Adds MultiNetworkPolicy definition to neubot manifest

* Removes static MultiNetworkPolicy definitions

These are now handled by a new template, and called from each experiment
manifest.

* Flattens arrays for experiment imports in system.jsonnet

The imports for experiments are now an array consisting of the experiment
DaemonSet definition and it's corresponsing MultiNetworkPolicy definition.

* Removes import of now nonexistent static MultiNetworkPolicies

* Casts endPort string to an int

The MultiNetworkPolicy CRD schema doesn't allow for that field to be a string,
unlike "port", which is odd, but whatever.

* Adds additional comments clarify custom iptables rules