Pins multi-networkpolicy-iptables DaemonSet to sandbox and staging
We need to test this configuration more in staging, and we don't want this to
block releasing k8s-support for other things in the meantime.

The MultiNetworkPolicy CRD definition will still be deployed to production,
along with the MultiNetworkPolicy instances for each experiment, as well as the
rbac configs for the DaemonSet. This just holds back the DaemonSet. It should
be perfectly find and safe for everything else to go to production. It will
exist in the production cluster, but will be unused until the DaemonSet is
there act on it.