Increases memory request and limit for multi-networkpolicy
I noticed that some multi-networkpolicy containers were getting OOM-killed.
Looking at typical memory usage of the container across all staging nodes, the
usage seems to hover right around 125MB for most of them, with usage sometimes
spiking up to the limit of 150MB and getting OOM-killed. This commit increases
the request to 150MB and sets the limit to 250MB. Based on monitoring data,
these should be more reasonable settings.

Additionally, I changed the label "name" label on the pod to "workload", which
is more consistent with how we label other pods. I removed the "app" label,
since we don't use that convention and it was, I believe, unused.