Add canary nodeSelector for multi-networkpolicy in mlab-oti
For sandbox and staging it's fine to deploy multi-networkpolicy everywhere, but
for production we want to do some canaries before deploying widely to be sure
that it works as intended and that it doesn't impact performance in any way.

Also, increase the memory limit to 500Mi. I have noticed that on physical
machines in sandbox multi-networkpolicy only uses around 30 or 40MB, but in
staging it uses more like 120MB. The only way I can account for this is that
staging has more nodes and perhaps the service somehow keeps track other nodes
for some reason.