Allow multi-networkpolicy to deploy to mlab-oti with canary nodeSelector (#913)
* Increases memory request and limit for multi-networkpolicy

I noticed that some multi-networkpolicy containers were getting OOM-killed.
Looking at typical memory usage of the container across all staging nodes, the
usage seems to hover right around 125MB for most of them, with usage sometimes
spiking up to the limit of 150MB and getting OOM-killed. This commit increases
the request to 150MB and sets the limit to 250MB. Based on monitoring data,
these should be more reasonable settings.

Additionally, I changed the label "name" label on the pod to "workload", which
is more consistent with how we label other pods. I removed the "app" label,
since we don't use that convention and it was, I believe, unused.

* Add canary nodeSelector for multi-networkpolicy in mlab-oti

For sandbox and staging it's fine to deploy multi-networkpolicy everywhere, but
for production we want to do some canaries before deploying widely to be sure
that it works as intended and that it doesn't impact performance in any way.

Also, increase the memory limit to 500Mi. I have noticed that on physical
machines in sandbox multi-networkpolicy only uses around 30 or 40MB, but in
staging it uses more like 120MB. The only way I can account for this is that
staging has more nodes and perhaps the service somehow keeps track other nodes
for some reason.

* Allows multi-networkpolicy DS to deploy to mlab-oti

For now it has a canary nodeSelector, so it will only to deploy to nodes that
we manually label with that nodeSelector.

* Uses versioned image for multi-networkpolicy DaemonSet