Add support for autojoined VMs (#919)
* Rought first stab at VMs using the Autojoin API for DNS

We are interested in the idea of having platform VMs use the Autojoin API to
decouple virtual sites and siteinfo. This is a first stab at testing that by
creating a new ndt-virtual-autojoin DaemonSet. This isn't expected to work
end-to-end yet, but is just test to see what breaks.

* Adds new DaemonSet ndt-virtual-autojoin

This is an experimental DaemonSet that tries to use the Autojoin API to handle
registering a DNS name. It is not expected to be fully, or even partially,
funcational, but is a test to see what breaks and doesn't. This is iteration

* Adds volume "autonode" to ExperimentNoIndex()

Also, fixes a couple of syntax errors in the ndt-virtual-autojoin DaemonSet

* Fixes volume name from "autojoin" to "autonode"

Also, fixes an error in env variable fieldRef specification.

* Fixes one last mount named 'autojoin' instead of 'autonode'

* Updates version of Heartbeat and uuid-annotator

These newer versions should have introduced support for certain flags being
able to read their values from local files (via flagx).

* Scopes conditional in ndt-virtual-autojoin DaemonSet for heartbeat

Previously, jsonnet was evaluating regular flags that I didn't mean to be part
of the conditional as part of the else.

* Sets UUIDAnnotator() param autojoin to default to false

* Creates autojoin-api-key Secret from GCS file

* Adds wildcard certificates for "mlab" autojoin clients

* Actually adds wildcard certificate for mlab autojoin nodes

A previous commit just added the selector to the ClusterIssuer, not the actual
Certificate object.

* Updates ndt-server to v0.23.0

This new version implements the flag -token.machine as a flagx StringFile,
which allow you specify a value prefixed with a @, and the rest a path to some
file.

* Corrects careless syntax issue with use of jsonnet std.split()

* Fixes a temporary debug change that got left behind

* Fixes autojoin-api-key secret key name

* Updates autojoin-register container image to v0.2.8

* Updates autojoin-register to version v0.2.9

This update allowed me to take advantage of the fact that this new version
supports flagx.StringFile for flags -iata, -ipv4 and -ipv6 to read the values
from files in the local filesystem.

* Updates autonode-register container image version to v0.2.11

This brings in the change that supports the -type and -uplink flags to the
register command.

* Changes -uplink flag of autonode-register to 7g

For a long time we have bene using 1g as the uplink speed for virtual machines,
even though GCP VMs are capable of 7g, according to Google documentation. This
just changes the value to be in line with reality.

* Changes autonode-register -iata flag to use right path

* Adds *.mlab.autojoin to production certificates

Production autojoined machines have a project ID of "autojoin" in their
hostname, not "mlab-oti". For M-Lab-managed VMs this domain would not be
included in the certificate. This change adds the wildcard certificate
*.mlab.autojoin.meaurement-lab.org, but only when the project is mlab-oti.