Create new system for implementing nftables rules in experiment pods (#922)
* Removes all MultiNetworkPolicy things

We are not going to use it. It broke production, and is abandoned anyway.

* Removes all MultiNetworkPolicy stuff + adds nftables rules

* Mount nftables-rules in /etc/nftables.d, not /etc

* Runs initContainer set-nftables-rules as user 0

* Runs set-nftables-rules initContainer in privileged mode

But tries to strip all capabilities but NET_ADMIN

* Adds missing comma to templates.jsonnet

* Fixes small typo in nftables.jsonnet

* Adds nftables rules for packet-test.. allows everything, for now

* Allows established/related traffic to experiment pods

* Removes nftables initContainer from no-index DaemonSets

No-index DaemonSets use hostNetwork=true, and we don't want to apply new
nftables rules to the default netns. Namely, we don't want to flush the
nftables rules, which breaks kube-proxy, flannel and probably other things.

* Fixed ct nftables rules + adds ports 9990-9999 everywhere

The ct state tracking rule was broken/invalid. Also, not every set of rules was
unconditionally letting monitoring traffic through.

* Adds nftables rules for the packet-test experiment