Adds an "fyi_project=mlab-oti" to NDT test volume doubled FYI alert (#938)
* Adds an "fyi_project=mlab-oti" to NDT test volume doubled FYI alert

Testing volume is far too variable in sandbox and staging, causing the FYI
alert for when NDT testing volume more than doubled to fire. This commit adds a
new fyi_project=mlab-oti label to this FYI alert, and the Alertmanage will
only match this label when it is the version running in mlab-oti.

* Rolls back accidental commit of k8s version upgrade. These changes were
supposed to be part of a future PR, not this one.